# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: COGS (Cost of Goods Sold) Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}/role:
    patch:
      tags: [Users]
      operationId: updateUserRole
      summary: Promote or demote a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppUser'
...
components:
...
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, role]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
          enum: [admin, user]
...
    AppUser:
      $ref: '#/components/schemas/AuthUser'
    UpdateUserRoleBody:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, user]
...